% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_con_gen.R
\name{uniform_con_gen}
\alias{uniform_con_gen}
\title{Uniform Confidence Bands for a Marginal Treatment Effect (MTE) Function .}
\usage{
uniform_con_gen(
  outcome_name = outcome_name,
  covariate_name = covariate_name,
  instrument_name = instrument_name,
  treatment_name = treatment_name,
  data = data,
  family = "probit",
  trim = 0.01,
  intercept = TRUE,
  se_type = "HC2",
  bw = bw,
  l_eval = l_eval,
  u_eval = u_eval,
  covariate_value = covariate_value,
  significance_level = 0.05
)
}
\arguments{
\item{outcome_name}{The name of the outcome}

\item{covariate_name}{The name of covariate}

\item{instrument_name}{The name of instrument variables}

\item{treatment_name}{The name of a treatment variable}

\item{data}{data.frame to be used for the estimation of MTE}

\item{family}{The distribution used for the calculation of the propensity score.
You can choose "probit" or "logit". The default choice is "probit".}

\item{trim}{To mitigate the effect of ill behavior of estimated propensity scores,
we set those estimated values less than or larger than the value of trim  as the value of the value of trim. The default choice is "0.01".}

\item{intercept}{'TRUE' or 'FLASE' option to decide whether you include an intercept
for estimating the propensity score or not. The default choice is 'TRUE'.}

\item{se_type}{The sort of standard error sought.
The options are "HC0", "HC1" (or "stata", the equivalent) , "HC2" (default), "HC3", or "classical".}

\item{bw}{The bandwidth used in estimating the nonparametric part of MTE}

\item{l_eval}{The minimum value of evaluation points for the estimated MTE}

\item{u_eval}{The maximum value of evaluation points for the estimated MTE}

\item{covariate_value}{A covariate value to plot the estimate MTE.}

\item{significance_level}{The significance level for the construction of uniform confidence band.}
}
\value{
A list that contains the following elements:
\item{Data}{A data frame that contains the following elements: \cr outcome, covariate, instrument variables, treatment and propensity score}
\item{Estimation}{A data frame of the following results: \cr evaluation point of MTE, estimated value of MTE, standard error of MTE, Lower bound of a confidence band , Upper bound of a confidence band}
\item{SE}{A standard error of each coeffecient of parametric parts of MTE}
\item{critical_value}{ A critical value corresponding to the siginificance level chosen}
\item{Plot}{A list that contains the ggplot elements for unifrom confidence band}
}
\description{
\code{uniform_con_gen} computes uniform confidence bands for a MTE function that is estimated by a semiparametric estimation method. For details, see Tsuda, Jin and Okui (2024+).
}
\examples{
\dontrun{
set.seed(1)
library(MASS)
SS<-2000
X1<-rnorm(n=SS,mean=0,sd=1)
X2<-rnorm(n=SS,mean=0,sd=1)
X3<-rnorm(n=SS,mean=0,sd=1)
X4<-rnorm(n=SS,mean=0,sd=1)
Z1<-rnorm(n=SS,mean=0,sd=1)
Z2<-rnorm(n=SS,mean=0,sd=1)
Z3<-rnorm(n=SS,mean=0,sd=1)
Z4<-rnorm(n=SS,mean=0,sd=1)
E0<-c(1,0.5,0.3)
E1<-c(0.5,1,0.8)
E2<-c(0.3,0.8,1)
E<-cbind(E0,E1,E2)
E4<-mvrnorm(n=SS,mu=c(0,0,0),Sigma=E)
U0<-E4[,1]
U1<-E4[,2]
V<-E4[,3]
Expl<-cbind(X1,X2,X3,X4)
Instr<-cbind(X1,X2,X3,X4,Z1,Z2,Z3,Z4)
PInstr<-cbind(Z1,Z2,Z3,Z4)
beta0<-c(0.5,0.1,-0.1,-0.5)
beta1<-c(0.8,0.4,-0.4,-0.8)
ganma0<-c(0.4,0.4,0.4,0.4,0.3,0.3,0.3,0.3)
D<-numeric(SS)
for(i in 1:SS){
  if ((ganma0 \%*\% (Instr[i,]))>V[i]){
    D[i]=1}
}
Y<-numeric(SS)
for(i in 1:SS){
  Y[i]<-D[i]*((beta1\%*\%Expl[i,])+U1[i])+(1-D[i])*((beta0\%*\%Expl[i,])+U0[i])
}
demo=as.data.frame(cbind(Y,Expl,PInstr,D))
Supp_P_hut_U<-0.85
Supp_P_hut_L<-0.15
O_SS<-1000
B2<-seq(Supp_P_hut_L,Supp_P_hut_U,length=O_SS)
covariate_value<-c(0,0,0,0)
Final_Result_Ex<-uniform_con_gen(outcome_name="Y",
                                 covariate_name=c("X1","X2","X3","X4"),
                                 instrument_name=c("Z1","Z2","Z3","Z4"),
                                 treatment_name= "D",
                                 data=demo,
                                 family="probit",
                                 trim=0.01,
                                 intercept=TRUE,
                                 se_type="HC2",
                                 bw=0.1,
                                 l_eval=0.15,
                                 u_eval=0.85,
                                 covariate_value=covariate_value,
                                 significance_level=0.05
)
}




}
\references{
Tsuda, Jin and Okui (2024+).
Uniform Confidence Band for Marginal Treatment Effect Function.
will be available on arXiv
}
